backend/requirements.txt
aniso8601==6.0.0
click==8.1.3
colorama==0.4.4
config==0.5.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug==2.1.2
